// src/app/core/enums/permissions/shop-permission.ts
export enum ShopPermission {
  CategoryAddUpdate = 101,
  CategoryRead = 102,

  ProductBaseAddUpdate = 201,
  ProductBaseRead = 202,

  ProductAddUpdate = 301,
  ProductRead = 302,

  AdCamaignAddUpdate = 401,
  AdCamaignDelete = 402,
  AdCamaignRead = 403,

  PromotionAddUpdate = 501,
  PromotionDelete = 502,
  PromotionRead = 503,
}